+++ src/DoctrineBridge/EntityMessage/EntityMessageCollectionInterface.php
@@ -9,13 +9,13 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Component\Messenger\Recorder;
+namespace Symfony\Bridge\Doctrine\EntityMessage;
 
 /**
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
  * @author Matthias Noback <matthiasnoback@gmail.com>
  */
-interface RecordedMessageCollectionInterface
+interface EntityMessageCollectionInterface
 {
     /**
      * Fetch recorded messages.
